Auto allocate and encode foreign types

+++ examples/plugin/src/lib.rs
@@ -7,10 +7,10 @@ extern "C" {
 }
 
 #[no_mangle]
-pub extern "C" fn add_number(a: i32) -> i32 {
+pub extern "C" fn add_number(a: i32) -> f32 {
     unsafe {
         print_number(a);
-        a + get_number(a)
+        a as f32 + get_number(a) as f32
     }
 }
